fix strings
An escaped backspace just before the closing quote would be recognized
as an escaped quote and the rest of your program becomes a giant string.

Fennel on its own doesn't really deal with escape sequences, but the
underlying Lua platform does.  We can't really know its version so we
just support the latest, 5.4.