remove keywords and operators
They're just symbols.

However, they are symbols with dots in them (at least `.` and `..` are).
Yet, they're not multisyms either.  Indeed, a symbol that begins or ends
on a dot is not a multisym.

I managed to get the starting dots working.  Unfortunately, I can't get
a symbol to only end on a dot without completely breaking
`multi_symbol`.  So, no `?.` for you. :(