properly support the nil safe opperator
Symbols with dots (like "?.") are multisyms, unless they begin or end on
a dot (again like "?.").

I tried keeping `symbol` a single token but it was impossible; "foo.bar"
would get tokenized as

  (symbol "foo.") (symbol "bar")

which couldn't be a `multi_symbol` as that requires the "." to be a
token on itself.  On the other hand, assigning lower lexical precedence
to the "ends on a dot" symbol would make it to never get chosen.  So I
gave up and made `symbol` a non-terminal.

Speaking of precedence, I was getting many conflicts between `symbol`
and `table`, `table_binding`, etc (because of the ":") so for the first
time, I added some `prec()`s instead of more conflicts.  Probably has
something to do with [Match Specificity][see-docs].

[see-docs]: https://tree-sitter.github.io/tree-sitter/creating-parsers#conflicting-tokens