Makefile: make test the default rule
Also changes the `tree-sitter` command to be specified by a variable.
I'm currently on a musl-based distro, and the `npm` `tree-sitter-cli`
package does not work because it downloads and tries to run a prebuilt
glibc-based binary. JavaScript devs, amIright?

Anyways, I'm using the `?=` syntax which isn't POSIX though I do
remember seeing a proposal to add it to POSIX.  I can't find it now
though.